pyproject.toml
[project]
name = "android"
version = "1.1.0"
description = ""
readme = "README.md"
authors = [
    { name = "Magniquick", email = "67406856+Magniquick@users.noreply.github.com" },
]
license = { text = "GPL-3.0-or-later" }
requires-python = "<4.0,>=3.10"
dependencies = [
    "requests<3.0.0,>=2.32.3",
    "cryptography<47.0.0,>=46.0.2",
    "lxml<7.0.0,>=6.0.2",
    "python-dotenv<2.0.0,>=1.0.1",
]

[dependency-groups]
dev = [
    "lxml-stubs>=0.5.1",
    "types-requests>=2.32.4.20250913",
]

